updated formatting
fixed format to fit in the dashboard and be included there. Need to still edit in order for it to display correctly.
